+++ src/index.ts
@@ -11,6 +11,7 @@ export * from "./interactive";
 export * from "./networkConfig";
 export * from "./networkStake";
 export * from "./networkParams";
+export * from "./signableMessage";
 export * from "./utils";
 export * from "./proxyProvider";
 export * from "./apiProvider";
@@ -18,6 +19,7 @@ export * from "./scArgumentsParser";
 export * from "./esdtHelpers";
 export * from "./token";
 
+export * from "./crypto";
 export * from "./walletcore";
 export * from "./nullSigner";