+++ src/walletcore/index.ts
@@ -9,4 +9,5 @@ export * from "./userWallet";
 export * from "./userKeys";
 export * from "./validatorKeys";
 export * from "./userSigner";
+export * from "./userVerifier";
 export * from "./validatorSigner";