+++ src/walletcore/userWallet.ts
@@ -1,22 +1,10 @@
-import * as errors from "../errors";
-import nacl from "tweetnacl";
 import { UserPublicKey, UserSecretKey } from "./userKeys";
-const crypto = require("crypto");
-import { v4 as uuidv4 } from "uuid";
-import scryptsy from "scryptsy";
-
-// In a future PR, improve versioning infrastructure for key-file objects in erdjs.
-const Version = 4;
-const CipherAlgorithm = "aes-128-ctr";
-const DigestAlgorithm = "sha256";
-const KeyDerivationFunction = "scrypt";
+import { EncryptedData, Encryptor, Decryptor, CipherAlgorithm, Version, KeyDerivationFunction, Randomness } from "../crypto";
+import {ScryptKeyDerivationParams} from "../crypto/derivationParams";
 
 export class UserWallet {
     private readonly publicKey: UserPublicKey;
-    private readonly randomness: Randomness;
-    private readonly ciphertext: Buffer;
-    private readonly mac: Buffer;
-    private readonly kdfparams: ScryptKeyDerivationParams;
+    private readonly encryptedData: EncryptedData;
 
     /**
      * Copied from: https://github.com/ElrondNetwork/elrond-core-js/blob/v1.28.0/src/account.js#L76
@@ -30,21 +18,9 @@ export class UserWallet {
      * passed through a password-based key derivation function (kdf).
      */
     constructor(secretKey: UserSecretKey, password: string, randomness: Randomness = new Randomness()) {
-        const kdParams = new ScryptKeyDerivationParams();
-        const derivedKey = UserWallet.generateDerivedKey(Buffer.from(password), randomness.salt, kdParams);
-        const derivedKeyFirstHalf = derivedKey.slice(0, 16);
-        const derivedKeySecondHalf = derivedKey.slice(16, 32);
-        const cipher = crypto.createCipheriv(CipherAlgorithm, derivedKeyFirstHalf, randomness.iv);
-
         const text = Buffer.concat([secretKey.valueOf(), secretKey.generatePublicKey().valueOf()]);
-        const ciphertext = Buffer.concat([cipher.update(text), cipher.final()]);
-        const mac = crypto.createHmac(DigestAlgorithm, derivedKeySecondHalf).update(ciphertext).digest();
-
+        this.encryptedData = Encryptor.encrypt(text, password, randomness);
         this.publicKey = secretKey.generatePublicKey();
-        this.randomness = randomness;
-        this.ciphertext = ciphertext;
-        this.mac = mac;
-        this.kdfparams = kdParams;
     }
 
     /**
@@ -58,24 +34,9 @@ export class UserWallet {
      * From an encrypted keyfile, given the password, loads the secret key and the public key.
      */
     static decryptSecretKey(keyFileObject: any, password: string): UserSecretKey {
-        const kdfparams = keyFileObject.crypto.kdfparams;
-        const salt = Buffer.from(kdfparams.salt, "hex");
-        const iv = Buffer.from(keyFileObject.crypto.cipherparams.iv, "hex");
-        const ciphertext = Buffer.from(keyFileObject.crypto.ciphertext, "hex");
-        const derivedKey = UserWallet.generateDerivedKey(Buffer.from(password), salt, kdfparams);
-        const derivedKeyFirstHalf = derivedKey.slice(0, 16);
-        const derivedKeySecondHalf = derivedKey.slice(16, 32);
-        
-        const computedMAC = crypto.createHmac(DigestAlgorithm, derivedKeySecondHalf).update(ciphertext).digest();
-        const actualMAC = keyFileObject.crypto.mac;
+        const encryptedData = UserWallet.edFromJSON(keyFileObject)
 
-        if (computedMAC.toString("hex") !== actualMAC) {
-            throw new errors.ErrWallet("MAC mismatch, possibly wrong password");
-        }
-
-        const decipher = crypto.createDecipheriv(keyFileObject.crypto.cipher, derivedKeyFirstHalf, iv);
-
-        let text = Buffer.concat([decipher.update(ciphertext), decipher.final()]);
+        let text = Decryptor.decrypt(encryptedData, password);
         while (text.length < 32) {
             let zeroPadding = Buffer.from([0x00]);
             text = Buffer.concat([zeroPadding, text]);
@@ -85,14 +46,23 @@ export class UserWallet {
         return new UserSecretKey(seed);
     }
 
-    /**
-     * Will take about:
-     *  - 80-90 ms in Node.js, on a i3-8100 CPU @ 3.60GHz
-     *  - 350-360 ms in browser (Firefox), on a i3-8100 CPU @ 3.60GHz
-     */
-    private static generateDerivedKey(password: Buffer, salt: Buffer, kdfparams: ScryptKeyDerivationParams): Buffer {
-        const derivedKey = scryptsy(password, salt, kdfparams.n, kdfparams.r, kdfparams.p, kdfparams.dklen);
-        return derivedKey;
+    static edFromJSON(keyfileObject: any): EncryptedData {
+        return new EncryptedData({
+            version: Version,
+            id: keyfileObject.id,
+            cipher: keyfileObject.crypto.cipher,
+            ciphertext: keyfileObject.crypto.ciphertext,
+            iv: keyfileObject.crypto.cipherparams.iv,
+            kdf: keyfileObject.crypto.kdf,
+            kdfparams: new ScryptKeyDerivationParams(
+              keyfileObject.crypto.kdfparams.n,
+              keyfileObject.crypto.kdfparams.r,
+              keyfileObject.crypto.kdfparams.p,
+              keyfileObject.crypto.kdfparams.dklen
+            ),
+            salt: keyfileObject.crypto.kdfparams.salt,
+            mac: keyfileObject.crypto.mac,
+        });
     }
 
     /**
@@ -101,54 +71,23 @@ export class UserWallet {
     toJSON(): any {
         return {
             version: Version,
-            id: this.randomness.id,
+            id: this.encryptedData.id,
             address: this.publicKey.hex(),
             bech32: this.publicKey.toAddress().toString(),
             crypto: {
-                ciphertext: this.ciphertext.toString("hex"),
-                cipherparams: { iv: this.randomness.iv.toString("hex") },
+                ciphertext: this.encryptedData.ciphertext,
+                cipherparams: { iv: this.encryptedData.iv },
                 cipher: CipherAlgorithm,
                 kdf: KeyDerivationFunction,
                 kdfparams: {
-                    dklen: this.kdfparams.dklen,
-                    salt: this.randomness.salt.toString("hex"),
-                    n: this.kdfparams.n,
-                    r: this.kdfparams.r,
-                    p: this.kdfparams.p
+                    dklen: this.encryptedData.kdfparams.dklen,
+                    salt: this.encryptedData.salt,
+                    n: this.encryptedData.kdfparams.n,
+                    r: this.encryptedData.kdfparams.r,
+                    p: this.encryptedData.kdfparams.p
                 },
-                mac: this.mac.toString("hex"),
+                mac: this.encryptedData.mac,
             }
         };
-    }
-}
-
-class ScryptKeyDerivationParams {
-    /**
-     * numIterations
-     */
-    n = 4096;
-
-    /**
-     * memFactor
-     */
-    r = 8;
-
-    /**
-     * pFactor
-     */
-    p = 1;
-
-    dklen = 32;
-}
-
-export class Randomness {
-    salt: Buffer;
-    iv: Buffer;
-    id: string;
-
-    constructor(init?: Partial<Randomness>) {
-        this.salt = init?.salt || Buffer.from(nacl.randomBytes(32));
-        this.iv = init?.iv || Buffer.from(nacl.randomBytes(16));
-        this.id = init?.id || uuidv4({ random: crypto.randomBytes(16) });
     }
 }